move LocalizationManager to internal package
