+++ src/main/java/dev/sirtimme/iuvo/internal/localization/LocalizationManager.java
@@ -1,4 +1,4 @@
-package dev.sirtimme.iuvo.api.localization;
+package dev.sirtimme.iuvo.internal.localization;
 
 import net.dv8tion.jda.api.interactions.DiscordLocale;
 import org.slf4j.Logger;
@@ -11,18 +11,8 @@ public class LocalizationManager {
     private static final Logger LOGGER = LoggerFactory.getLogger(LocalizationManager.class);
     private static final Map<DiscordLocale, ResourceBundle> bundles = new HashMap<>();
 
-    public static void addBundles(final String baseName, final DiscordLocale... discordLocales) {
-        for (final var discordLocale : discordLocales) {
-            final var locale = discordLocale.toLocale();
-            try {
-                final var userBundle = ResourceBundle.getBundle(baseName, locale);
-                bundles.put(discordLocale, userBundle);
-
-                LOGGER.info("Loaded bundle for locale '{}'", locale);
-            } catch (final MissingResourceException error) {
-                LOGGER.warn("Loading of bundles for locale '{}' failed: {}", locale, error.getMessage());
-            }
-        }
+    static {
+        addBundles(DiscordLocale.ENGLISH_US, DiscordLocale.GERMAN);
     }
 
     public static String getResponse(final String key, final DiscordLocale locale, final Object... values) {
@@ -30,6 +20,20 @@ public class LocalizationManager {
         final var template = new MessageFormat(bundle.getString(key));
 
         return template.format(values);
+    }
+
+    private static void addBundles(final DiscordLocale... discordLocales) {
+        for (final var discordLocale : discordLocales) {
+            final var locale = discordLocale.toLocale();
+            try {
+                final var userBundle = ResourceBundle.getBundle("localization/responses", locale);
+                bundles.put(discordLocale, userBundle);
+
+                LOGGER.info("Loaded bundle for locale '{}'", locale);
+            } catch (final MissingResourceException error) {
+                LOGGER.warn("Loading of bundles for locale '{}' failed: {}", locale, error.getMessage());
+            }
+        }
     }
 
     private static ResourceBundle getBundle(final DiscordLocale locale) {
