+++ src/main/java/dev/sirtimme/iuvo/internal/precondition/IsComponentAuthor.java
@@ -3,7 +3,7 @@ package dev.sirtimme.iuvo.internal.precondition;
 import dev.sirtimme.iuvo.api.precondition.IPrecondition;
 import net.dv8tion.jda.api.events.interaction.component.GenericComponentInteractionCreateEvent;
 
-import static dev.sirtimme.iuvo.api.localization.LocalizationManager.getResponse;
+import static dev.sirtimme.iuvo.internal.localization.LocalizationManager.getResponse;
 
 public class IsComponentAuthor implements IPrecondition<GenericComponentInteractionCreateEvent> {
     @Override
@@ -11,7 +11,7 @@ public class IsComponentAuthor implements IPrecondition<GenericComponentInteract
         final var authorId = event.getComponentId().split(":")[0];
 
         if (!event.getUser().getId().equals(authorId)) {
-            event.reply(getResponse("iuvo.error.notComponentAuthor", event.getUserLocale())).setEphemeral(true).queue();
+            event.reply(getResponse("iuvo.precondition.notComponentAuthor.invalid", event.getUserLocale())).setEphemeral(true).queue();
             return false;
         }
