+++ src/main/java/dev/sirtimme/iuvo/internal/precondition/IsRegistered.java
@@ -5,7 +5,7 @@ import dev.sirtimme.iuvo.api.precondition.IPrecondition;
 import dev.sirtimme.iuvo.api.repository.Repository;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 
-import static dev.sirtimme.iuvo.api.localization.LocalizationManager.getResponse;
+import static dev.sirtimme.iuvo.internal.localization.LocalizationManager.getResponse;
 
 public class IsRegistered implements IPrecondition<SlashCommandInteractionEvent> {
     private final Repository<? extends IEntity> repository;
@@ -19,7 +19,7 @@ public class IsRegistered implements IPrecondition<SlashCommandInteractionEvent>
         final var user = repository.get(event.getUser().getIdLong());
 
         if (user == null) {
-            event.reply(getResponse("iuvo.error.notRegistered", event.getUserLocale())).queue();
+            event.reply(getResponse("iuvo.precondition.notRegistered.invalid", event.getUserLocale())).queue();
             return false;
         }
