+++ src/main/resources/localization/responses_de.properties
@@ -1,3 +1,3 @@
-iuvo.error.notOwner=Dieser Befehl kann nur von dem Besitzer der App ausgeführt werden
-iuvo.error.notRegistered=Du bist nicht registriert. Bitte benutze zuerst `/register`
-iuvo.error.notComponentAuthor=Du bist nicht der Autor des initialen Befehls
+iuvo.precondition.notOwner.invalid=Dieser Befehl kann nur von dem Besitzer der App ausgeführt werden
+iuvo.precondition.notRegistered.invalid=Du bist nicht registriert. Bitte benutze zuerst `/registrieren`
+iuvo.precondition.notComponentAuthor.invalid=Du bist nicht der Autor des initialen Befehls
