+++ src/main/resources/localization/responses_en_US.properties
@@ -1,3 +1,3 @@
-iuvo.error.notOwner=This command can only be executed by the owner of this app
-iuvo.error.notRegistered=You are not registered. Please use `/register` first
-iuvo.error.notComponentAuthor=You are not the author of the initial command
+iuvo.precondition.notOwner.invalid=This command can only be executed by the owner of this app
+iuvo.precondition.notRegistered.invalid=You are not registered. Please use `/register` first
+iuvo.precondition.notComponentAuthor.invalid=You are not the author of the initial command
